Move translator into helper module, output missing key info.
